откорректирован функционал по загрузке и отправке кортинки

+++ src/main/java/com/skypro/resale/repository/AvatarRepository.java
@@ -4,5 +4,5 @@ import com.skypro.resale.model.Avatar;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface AvatarRepository extends JpaRepository<Avatar, Long> {
+public interface AvatarRepository extends JpaRepository<Avatar, Integer> {
 }
